requirements.txt
apipkg==1.4
appdirs==1.4.0
beautifulsoup4==4.6.0
colorama==0.3.7
execnet==1.4.1
Faker==0.7.11
google-api-python-client==1.6.4
httplib2==0.10.3
oauth2client==4.1.2
packaging==16.8
py==1.4.32
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyparsing==2.1.10
pytest==3.0.6
pytest-html==1.14.2
pytest-metadata==1.5.1
pytest-rerunfailures==2.1.0
pytest-xdist==1.15.0
python-dateutil==2.6.1
rsa==3.4.2
selenium==3.7.0
six==1.11.0
uritemplate==3.0.0